+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.31.0 (Unreleased)
+
+ENHANCEMENTS:
+
+* Resource: `tencentcloud_tcaplus_idl` add new attribute `zone_id`. 
+
 ## 1.30.5 (March 19, 2020)
 
 BUG FIXES: